requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.0.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
